[백병화] 02-14 lottie_text_widget 추가 - Rectangle, LottieText, LottieTextWidget 클래스 추가

+++ lib/vm_sdk/impl/ffmpeg_argument_generator.dart
@@ -135,8 +135,8 @@ Future<GenerateArgumentResponse> generateVideoRenderArgument(
 
   // ADD TITLE
 
-  exportedTitle.width = (exportedTitle.width * 1.25).floor();
-  exportedTitle.height = (exportedTitle.height * 1.25).floor();
+  exportedTitle.width = ((exportedTitle.width * 1.25).floor()).toDouble();
+  exportedTitle.height = ((exportedTitle.height * 1.25).floor()).toDouble();
 
   final double startPosY = (videoHeight / 2) - (exportedTitle.height / 2);
